Moved InfluxTimeUnit to base Influx package
