+++ api/metrics-api-jaxrs/src/test/java/org/hawkular/metrics/api/jaxrs/influx/InfluxTimeUnitTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.api.jaxrs.influx.query.parse.definition;
+package org.hawkular.metrics.api.jaxrs.influx;
 
 import static org.junit.Assert.assertEquals;
